+++ benchmark/src/main.rs
@@ -1,4 +1,4 @@
-#[allow(dead_code)]
+#[allow(dead_code, unused_imports)]
 mod infra;
 #[allow(dead_code, unused_imports)]
 mod loadgen;
@@ -12,6 +12,7 @@ use anyhow::{Ok, Result};
 #[tokio::main]
 async fn main() -> Result<()> {
     // 1. Set up Postgres and Redis
+
     // 1.1 check health
 
     // 2. Build test-keys and api-keys
